src/external/HMAR/requirements.txt
# Copyright (c) 2025, NVIDIA Corporation. All rights reserved.
#
# This work is made available under the NVIDIA One-Way Noncommercial License v1 (NSCLv1).
# To view a copy of this license, please refer to LICENSE

Pillow
huggingface_hub
numpy
pytz
transformers
typed-argument-parser
typing
tqdm
scipy
requests
flash-attn
wandb
tensorflow
torchvision
torchtext
prettytable
hydra-core
triton
protobuf
tqdm